combining services, client and server code to one sub package
