+++ src/services/register_services_test.go
@@ -1,4 +1,4 @@
-package server
+package services
 
 import (
 	"testing"
